+++ CHANGELOG.md
@@ -48,6 +48,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - pykdtree
   - pyogrio
   - contourpy
+  - sunpy
 
 - Explicit Pip Packages
   - lxml